requirements.txt
# Core Dependencies
crawl4ai>=0.7.4
openai>=1.40.0
supabase>=2.7.0
pydantic>=2.8.0
python-dotenv>=1.0.0

# Web Scraping & HTTP
aiohttp>=3.9.0
aiofiles>=24.1.0
requests>=2.31.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Development Tools
black>=24.8.0
mypy>=1.11.0
ruff>=0.6.0

# Utility
pyyaml>=6.0.0
tqdm>=4.66.0